Eval overrides before passing to shading network
For large shading networks with overrides the evals would run per
node per override which adds up. This way the evals are just run
once per override and passed down. While this results in a few lines
of code duplication in PBRTgeo and PBRTwrangler, it cuts down on the
evals.