Fixed Sphere/Cylinder Area Lights.
These both had a ReverseOrientation and a Scale(1,1,-1) to match
Houdini's UVs. Unfortunately there appears to be an issue in pbrt-v4
when using a Scale(1,1,-1) results in illumination being on the
opposite side of the emissive surface. (One side bright, but the dark
side would be the one illuminating other surfaces.) Consistent lights
are more important that matching UVs.

The standard Sphere primitives were also updated to take a new render
property to allow them to match Houdini's parametric uvs or not. It
defaults to matching, but have the ability to turn this off so you
can use standard spheres as area lights as well.